Refactor createIssue
