+++ Source/Server/TeamTasker.Server.Application/Interfaces/IIssueService.cs
@@ -6,10 +6,9 @@ namespace TeamTasker.Server.Domain.Interfaces
 {
     public interface IIssueService
     {
-        void CreateIssue(CreateIssueDto issueDto, string email);
         IEnumerable<ReadIssueDto> GetAllIssues();
         ReadIssueDto GetIssue(int id);
-        void AddIssueToProject(AddIssueToProjectDto issueDto);
+        void AddIssueToProject(AddIssueToProjectDto issueDto); //Probably unnecessary 
 
         //IEnumerable<GetCompletedIssueDto> GetCompletedIssue();
         //IEnumerable<GetCompletedIssueDto> GetNotCompletedIssue();
